Move Address class from top-level file into CheckoutController.
