+++ paymentsheet/src/main/java/com/stripe/android/checkout/CheckoutCollectedDetails.kt
@@ -1,6 +1,7 @@
 package com.stripe.android.checkout
 
 import android.os.Parcelable
+import com.stripe.android.checkout.CheckoutController.Address
 import com.stripe.android.paymentelement.CheckoutSessionPreview
 import kotlinx.parcelize.Parcelize
